Completed basic use-case

+++ cs544-202403-project/src/main/java/edu/miu/cs/cs544/controller/MemberController.java
@@ -9,5 +9,5 @@ import edu.miu.cs.cs544.service.contract.MemberPayload;
 
 @RestController
 @RequestMapping("/members")
-public class MemberController extends BaseReadWriteController<MemberPayload, Member, Integer> {
+public class MemberController extends BaseReadWriteController<MemberPayload, Member, Long> {
 }
