+++ cs544-202403-project/src/main/java/edu/miu/cs/cs544/service/MemberServiceImpl.java
@@ -1,12 +1,10 @@
 package edu.miu.cs.cs544.service;
 
-import org.springframework.stereotype.Service;
-
 import edu.miu.common.service.BaseReadWriteServiceImpl;
 import edu.miu.cs.cs544.domain.Member;
 import edu.miu.cs.cs544.service.contract.MemberPayload;
+import org.springframework.stereotype.Service;
 
 @Service
-public class MemberServiceImpl extends BaseReadWriteServiceImpl<MemberPayload, Member, Integer> implements MemberService {
-	
+public class MemberServiceImpl extends BaseReadWriteServiceImpl<MemberPayload, Member, Long> implements MemberService {
 }
